+++ CHANGELOG.md
@@ -1,4 +1,26 @@
-# Change Log
+# CHANGELOG
+
+## [0.1.6]
+
+- Add YouTube Information Retrieval from URL or Video ID
+
+## [0.1.5]
+
+- Added DeepL translation usage status bar
+
+## [0.1.4]
+
+- Add DeepL Translation command
+
+## [0.1.3]
+
+## [0.1.2]
+
+- Add Ollama Summarization command
+
+## [0.1.1]
+
+- Add Status Bar
 
 ## [0.1.0]